# spring-client/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: legal@coindcx.com
  description: "Welcome to the CoinDCX API!\n\nThe base URL for the API calls is `https://api.coindcx.com`,\
    \ base URL for some public endpoint is `https://public.coindcx.com`.\n\n## Important\
    \ Legal Notice\n\nThese API License Terms and Conditions govern the use of any\
    \ 'Market Data' and Application Programming Interface (API) of CoinDCX. \nBy accessing\
    \ any CoinDCX API, Market Data and/or any other information, service, or feature,\
    \ you shall be bound by these Terms.\n\n## Key Definitions\n\n- **Market Data**:\
    \ All data related to trading activity on CoinDCX platforms, including prices,\
    \ quantities of orders and transactions\n- **CoinDCX API**: Public API and/or\
    \ any data, information related to trading activity made available through https://docs.coindcx.com/\n\
    - **Algorithmic Trading**: Method using computer programs following defined instructions\
    \ or algorithms to place trades\n- **High-Frequency Trading**: Type of algorithmic\
    \ trading with high speeds, high turnover rates, and high order-to-trade ratios\n\
    \n## License Grant\n\nUsers are granted a **non-exclusive, non-transferable, non-assignable,\
    \ non-sublicensable, revocable, restricted license** for usage only.\n\nUsers\
    \ must NOT:\n- Alter, manipulate, or misrepresent any CoinDCX API or Market Data\n\
...
    \ a timestamp of when it was generated\n"
  license:
...
  title: CoinDCX API
...
servers:
...
security:
- ApiKeyAuth: []
  SignatureAuth: []
...
paths:
...
  /exchange/v1/orders/create:
    post:
      description: Place a new order on the exchange. Maximum of 25 open orders at
        a time for one specific market.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_exchange_v1_orders_create_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_exchange_v1_orders_create_post_200_response"
          description: Order created successfully
      security:
      - ApiKeyAuth: []
      summary: Create new order
      tags:
      - Order
      x-content-type: application/json
      x-accepts:
      - application/json
...
  /exchange/v1/orders/status:
    post:
      description: Fetch status of any order
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_exchange_v1_orders_status_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_exchange_v1_orders_create_post_200_response_orders_inner"
          description: Successful response
      security:
      - ApiKeyAuth: []
      summary: Get order status
      tags:
      - Order
      x-content-type: application/json
      x-accepts:
      - application/json
...
  /exchange/v1/orders/cancel:
    post:
      description: Cancel an active order
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_exchange_v1_orders_cancel_post_request"
        required: true
      responses:
        "200":
          description: Order cancelled successfully
      security:
      - ApiKeyAuth: []
      summary: Cancel order
      tags:
      - Order
      x-content-type: application/json
      x-accepts:
      - application/json
  /exchange/v1/orders/edit:
    post:
      description: Edit the price of an active order
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_exchange_v1_orders_edit_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: Order edited successfully
      security:
      - ApiKeyAuth: []
      summary: Edit order price
      tags:
      - Order
      x-content-type: application/json
      x-accepts:
      - application/json
...
  /exchange/v1/derivatives/futures/orders/create:
    post:
      description: Place a new futures order
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_exchange_v1_derivatives_futures_orders_create_post_request"
        required: true
      responses:
        "200":
          description: Order created successfully
        "400":
          description: "Bad request (insufficient funds, invalid price, etc.)"
        "422":
          description: Validation error
      security:
      - ApiKeyAuth: []
      - SignatureAuth: []
      summary: Create Futures Order
      tags:
      - Futures
      x-content-type: application/json
      x-accepts:
      - application/json
  /exchange/v1/derivatives/futures/orders/cancel:
    post:
      description: Cancel an existing futures order by ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_exchange_v1_derivatives_futures_orders_cancel_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_exchange_v1_derivatives_futures_orders_cancel_post_200_response"
          description: Order cancelled successfully
      security:
      - ApiKeyAuth: []
      - SignatureAuth: []
      summary: Cancel Futures Order
      tags:
      - Futures
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
...
    _exchange_v1_orders_create_post_request:
      properties:
        market:
          description: The trading pair
          example: SNTBTC
          type: string
        total_quantity:
          description: Quantity to trade
          example: 1.101
          type: number
        price_per_unit:
          description: Price per unit (not required for market order)
          example: 0.082
          type: number
        side:
          description: Specify buy or sell
          enum:
          - buy
          - sell
          type: string
        order_type:
          description: Order Type
          enum:
          - market_order
          - limit_order
          type: string
        client_order_id:
          description: Client order id of the order
          example: 2022.02.14-btcinr_1
          type: string
        timestamp:
          description: Current timestamp in milliseconds
          format: int64
          type: integer
      required:
      - market
      - order_type
      - side
      - timestamp
      - total_quantity
      type: object
    _exchange_v1_orders_create_post_200_response_orders_inner:
      example:
        side: side
        fee: 6.027456183070403
        created_at: 2000-01-23T04:56:07.000+00:00
        avg_price: 5.637376656633329
        remaining_quantity: 5.962133916683182
        market: market
        fee_amount: 0.8008281904610115
        updated_at: 2000-01-23T04:56:07.000+00:00
        total_quantity: 1.4658129805029452
        id: id
        price_per_unit: 2.3021358869347655
        order_type: order_type
        status: status
      properties:
        id:
          type: string
        market:
          type: string
        order_type:
          type: string
        side:
          type: string
        status:
          type: string
        fee_amount:
          type: number
        fee:
          type: number
        total_quantity:
          type: number
        remaining_quantity:
          type: number
        avg_price:
          type: number
        price_per_unit:
          type: number
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
      type: object
    _exchange_v1_orders_create_post_200_response:
      example:
        orders:
        - side: side
          fee: 6.027456183070403
          created_at: 2000-01-23T04:56:07.000+00:00
          avg_price: 5.637376656633329
          remaining_quantity: 5.962133916683182
          market: market
          fee_amount: 0.8008281904610115
          updated_at: 2000-01-23T04:56:07.000+00:00
          total_quantity: 1.4658129805029452
          id: id
          price_per_unit: 2.3021358869347655
          order_type: order_type
          status: status
        - side: side
          fee: 6.027456183070403
          created_at: 2000-01-23T04:56:07.000+00:00
          avg_price: 5.637376656633329
          remaining_quantity: 5.962133916683182
          market: market
          fee_amount: 0.8008281904610115
          updated_at: 2000-01-23T04:56:07.000+00:00
          total_quantity: 1.4658129805029452
          id: id
          price_per_unit: 2.3021358869347655
          order_type: order_type
          status: status
      properties:
        orders:
          items:
            $ref: "#/components/schemas/_exchange_v1_orders_create_post_200_response_orders_inner"
          type: array
      type: object
...
    _exchange_v1_orders_status_post_request:
      properties:
        id:
          description: The ID of the order
          example: ead19992-43fd-11e8-b027-bb815bcb14ed
          type: string
        client_order_id:
          description: The Client Order ID of the order
          example: 2022.02.14-btcinr_1
          type: string
        timestamp:
          format: int64
          type: integer
      required:
      - timestamp
      type: object
...
    _exchange_v1_orders_cancel_post_request:
      properties:
        id:
          description: The ID of the order
          example: ead19992-43fd-11e8-b027-bb815bcb14ed
          type: string
        timestamp:
          format: int64
          type: integer
      required:
      - id
      - timestamp
      type: object
    _exchange_v1_orders_edit_post_request:
      properties:
        id:
          description: The ID of the order
          example: ead19992-43fd-11e8-b027-bb815bcb14ed
          type: string
        price_per_unit:
          description: New price for the order
          example: 123.45
          type: number
        timestamp:
          format: int64
          type: integer
      required:
      - id
      - price_per_unit
      - timestamp
      type: object
...
    _exchange_v1_derivatives_futures_orders_create_post_request:
      properties:
        timestamp:
          description: EPOCH timestamp in milliseconds. Orders with delay > 10 seconds
            will be rejected
          format: int64
          type: integer
        side:
          description: Order side
          enum:
          - buy
          - sell
          type: string
        pair:
          description: Pair name (format B-ETH_USDT)
          example: B-ETH_USDT
          type: string
        order_type:
          description: Order type
          enum:
          - market
          - limit
          - stop_limit
          - stop_market
          - take_profit_limit
          - take_profit_market
          type: string
        price:
          description: "Order Price (limit price for limit, stop limit, and take profit\
            \ limit orders). NULL for market orders"
          type: number
        stop_price:
          description: "Stop Price (for stop_limit, stop_market, take_profit_limit,\
            \ take_profit_market orders). Trigger price"
          type: number
        total_quantity:
          description: Order total quantity
          type: number
        leverage:
          description: Leverage at which to take position. Should match position leverage
          type: number
        notification:
          description: Email notification preference
          enum:
          - no_notification
          - email_notification
          type: string
        time_in_force:
          default: good_till_cancel
          description: Time in force (null for market orders)
          enum:
          - good_till_cancel
          - fill_or_kill
          - immediate_or_cancel
          type: string
        hidden:
          description: Not supported at the moment
          type: boolean
        post_only:
          description: Not supported at the moment
          type: boolean
        margin_currency_short_name:
          default: USDT
          description: Futures margin mode
          enum:
          - INR
          - USDT
          type: string
        position_margin_type:
          description: Position margin type
          enum:
          - isolated
          - crossed
          type: string
        take_profit_price:
          description: "Take profit trigger price (only for market_order, limit_order)"
          type: number
        stop_loss_price:
          description: "Stop loss trigger price (only for market_order, limit_order)"
          type: number
      required:
      - notification
      - order_type
      - pair
      - side
      - timestamp
      - total_quantity
      type: object
    _exchange_v1_derivatives_futures_orders_cancel_post_request:
      properties:
        timestamp:
          description: EPOCH timestamp in milliseconds
          format: int64
          type: integer
        id:
          description: Order ID
          type: string
      required:
      - id
      - timestamp
      type: object
    _exchange_v1_derivatives_futures_orders_cancel_post_200_response:
      example:
        code: 6
        message: message
        status: 0
      properties:
        message:
          type: string
        status:
          type: integer
        code:
          type: integer
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: API key for authentication
      in: header
      name: X-AUTH-APIKEY
      type: apiKey
    SignatureAuth:
      description: HMAC-SHA256 signature of the request payload
      in: header
      name: X-AUTH-SIGNATURE
      type: apiKey